refactore: separate to constants and types files
