+++ client/features/playground/types.ts
@@ -1,0 +1,2 @@
+export type Block = { id: number; arg: (Block | string)[] };
+export type BLOCK = { id: number; contents: string[] };
